+++ korge-fleks/src/commonMain/kotlin/korlibs/korge/fleks/components/SpawnerComponent.kt
@@ -17,7 +17,7 @@ data class SpawnerComponent(
     var numberOfObjects: Int = 1,                  // The spawner will generate this number of object when triggered after interval time
     var interval: Int = 1,                         // 0 - disabled, 1 - every frame, 2 - every second frame, 3 - every third frame,...
     var timeVariation: Int = 0,                    // 0 - no variation, 1 - one frame variation, 2 - two frames variation, ...
-    var positionVariation: Float = 0.0f,           // variation radius where objects will be spawned - 0.0 = no variation
+    var positionVariation: Double = 0.0,           // variation radius where objects will be spawned - 0.0 = no variation
     var newEntity: Entity = invalidEntity,         // If spawner shall take a specific entity for spawning it can be set here
     var config: Identifier = nothing,             // Name of configuration which is used in below function to configure the new entity
     var function: Identifier = nothing,           // Name of function which configures the new entity
